Move sort to heavy-lifting class
