+++ src/InMemoryRepository.php
@@ -214,22 +214,8 @@ class InMemoryRepository implements ObjectRepository, Selectable
     {
         $expr = $criteria->getWhereExpression();
 
-        $matcher = new CriteriaEvaluator($this->getClassName(), $this->managedEntities);
-        $entities = $matcher->match($expr);
-
-        if ($orderings = $criteria->getOrderings()) {
-            $entities = $this->sortResults($entities, $orderings);
-        }
-
-        if ($offset = $criteria->getFirstResult()) {
-            $entities = array_slice($entities, $offset);
-        }
-
-        if ($limit = $criteria->getMaxResults()) {
-            $entities = array_slice($entities, 0, $limit);
-        }
-
-        return $entities;
+        return (new CriteriaEvaluator($this->getClassName(), $this->managedEntities))
+            ->evaluate($criteria);
     }
 
     /**
@@ -264,56 +250,6 @@ class InMemoryRepository implements ObjectRepository, Selectable
         $propVal = $rp->getValue($entity);
         // $rp->setAccessible($isAccessible);
         return $propVal;
-    }
-
-    /**
-     * @param Entity[] $results
-     * @param array<array-key, string> $orderBy
-     * @return Entity[]
-     */
-    private function sortResults(array $results, array $orderBy): array
-    {
-        // WARNING: this has the potential to get quite slow due to the use
-        // of reflection on TWO entities on every single comparision. On
-        // typical test datasets this is unlikely to be a major issue, but
-        // it's quite inefficient and should be re-examined in the future.
-        /**
-         * @param Entity $a
-         * @param Entity $b
-         */
-        usort($results, function ($a, $b) use ($orderBy) {
-            foreach ($orderBy as $propName => $direction) {
-                $v1 = $this->getValueOfProperty($a, $propName);
-                $v2 = $this->getValueOfProperty($b, $propName);
-                if ($v1 === $v2) {
-                    // Don't return 0 - that's only safe if on the last
-                    // property in the sorting criteria
-                    continue;
-                }
-                if ($direction === Criteria::ASC) {
-                    if ($v1 > $v2) {
-                        return 1;
-                    } else {
-                        return -1;
-                    }
-                } elseif ($direction === Criteria::DESC) {
-                    if ($v1 < $v2) {
-                        return 1;
-                    } else {
-                        return -1;
-                    }
-                } else {
-                    // Should be unreachable
-                    // @codeCoverageIgnoreStart
-                    throw ORMException::invalidOrientation($this->getClassName(), $propName);
-                    // @codeCoverageIgnoreEnd
-                }
-            }
-            // If all loops have exited without returning a comparision
-            // result, all of the sort properties should be equal.
-            return 0;
-        });
-        return $results;
     }
 
     /**
